refactor(frontend): split desktop workspace shell responsibilities
